+++ whatsnew.md
@@ -3,6 +3,8 @@ Version 0.1.2
 * Fixed bug in `FitHarn._check_divergence`
 * Added dependency on astunparse to fix bug where exporter could not handle complex assignments
 * Pretrained initializer now only requires the path to the deploy zipfile. It can figure out which files in the deployment are the weights.
+* Can now look up images by filename
+* Can now delete categories by category name
 
 
 Version 0.1.1